//Circuit.cpp --194-236------- Changed the setters for fuzz and vol to default to the max and min values rather than having a default value.
--247-252-----
Added the setParams method to set the parameters to the arguements

//Circuit.h
--7-8-----
Added the namespaces constants to the header for access to global "Project Settings"

--52-60----
Made the setters and getters "protected" so only the simulation class can call it.

///PluginEditor.cpp////
--79-----
--96-101-----
--111-112----
--139-141----
Added the gain Knob to the UI and changed the positioning and size of the other elements to fit

//PluginProcessor.cpp
--21-26----
Added the gain param for controlling input gain (work in progress)

--148-154-----
Added a check for param changes before implementing the setParams method (saves some CPU usage so the system only refreshes system matrices when a param actually changes)

--163-167----
Implemented the new pointer based processSample instead of the Eigen Matrix based process buffer. This processes each sample using the pointer to the audio data.

//ProjectSettings.h
Added this header file which contains a namespace containing ALL global settings for the project.
This should be referenced throughout the project to avoid any "magic numbers" in the code.

//Simulation.cpp
--122-124------
Changed the processSample method so that it accepts the float*  _channelData pointer as arguement to better align with how the processBlock method in JUCE works.

--209-213-------
Added the setParams method as an override of the Circuit setParams method.
This method calls the Circuit setParams and then refreshes the system matrix with updated parameter vals.