//Circuit.cpp -10------ changed the value T to samplePeriod for better clarity (line 79 in circuit.h)
--193-230---
Changed the rules around setting the fuzz and vol params so that they are always set to some value related to the control increment

//Circuit.h/////
--50------
Added the macro to ensure alignment of eigen members

--68-76---
Made these members protected instead of private so that the Simulation class has access.

//ProjectSettings.h///
Added more defaults and constants to the settings page.

//Simulation.cpp
--64-65-----
Float* pointer used as input signal for steady state phase

--72-//-----
changed MM to hanWin to clarify that this is refering to the hanning window length in samples.

--111-114---
Deallocate and clean-up the pointer

//Simulation.h
--41------
Added the macro to align eigen members

--54-----
Updated the comments for the new hanWin integer

--!61----
Removed the powerUpTimeVector

//MyLookAndFeel.class
Added the new MyLookAndFeel class which extends the LookAndFeel_V3, overriding the drawRotarySlider method to allow for use of the "film strip" stored in the binary data to be used as the knob image rather than base class's own vector drawing of the knob.

//PluginEditor.cpp
--34-41----
Changed the valueChanged method to include a check to see if the rotary is being actively held. This lets the slider be automated by the host DAW if the user is not actively trying to control the parameter

--69-77-----
Changed the updateSliderPos method to include a check to see if the user is trying to control the parameter.  This means that if the parameter is being automated by the host, and the user then tries to change the parameter themselves, the user's changes will take priority over the automation.

--94---
Added initialiser for the input signal label

--116-135-----
The knobs now no longer have an editable textbox, instead a popup box showing the value appears when the user changes the parameter.

--141-150--
Changed the text colour to white

--152-154---
Added the inputSignalLabel to the view

--157-160--
Set the look and feel of the 3 rotary sliders to myLookAndFeel to implement the film strip knob.

--170-
Start the timer for updating the input signal label

--177-180---
timerCallback method updates the input signal label with the current input every P_TIMER_FREQ

--185-189---
Changed the backgorund colour of the plugin window.

--192--
Set text colour white

--203---
Position the input signal label.

//PluginEditor.h
--21---
Extend the Timer class

--30---
declare the timerCallback method

--36-37---
Declare the label and double for storing input signal

--40--
Scoped pointer for myLookAndFeel

--44--
Scoped pointer for the sliders

//PluginProcessor.cpp
--26---
Default the fuzz to min

--34-40---
Initialise the samplerate
Initialise the parameter smoothing coeff

--166-177--
Check if parameters have changed, if so, update.

--193--
Apply the gain to the channelData

--196--
Scale the input

--198--
update the currentInput double for the input signal label

--202---
Process the data

--204---
Scale the output

--209-220--
Scale the input to the system by the factor INPUT_SCALAR,

If the scaled input is still above the clipping point then hard clip the signal, this stops the simulation from crashing due to too high a signal being input to the system.

--234--
update the fuzzVal with the smoothed Fuzz param

--238-246--
Added parameter smoothing methods for vol and fuzz

//PluginProcessor.h
--30--
Scoped pointer for the simulation instance, this fixed the crash where improper memory allocation meant that the plugin didn't work with some hosts and completely crashed others

--33-35--
Changed all the parameters to scoped pointers to avoid memory leaks

--40-41--
double for sampeRate and smoothing coefficient

--45-46--
Declare the parameter smoothing methods

--85--
currentInput is the value to be displayed on the inputSignal Label

--89--
Declare the timerCallback override

--92--
declare the inputScaling method

//BinaryData.class
This class is auto generated by JUCE to store the png of the Knob Filmstrip as binary data, for use within the MyLookAndFeel class